During sync stage we are using delta cache(for rollback and for commit modes). If during rollback we "revert removing of MultisigEntry", we will add it to addedElements set. If after that during commit mode we will remove this MultisigEntry again, we will remove it from addedElements.
When we calculate hash of cache, we are using partial tree. In this tree we apply add and remove operation.
But during sync we can miss remove operation. Harvester of block removed this MultisigEntry from his cache, and processed it in partial tree correct, when in sync mode we can not process remove operation and our hash will different.