UtUpdater locks(reader) cache by self. When transaction is failed, m_failedTransactionSink will notify about status of this transaction. During notification we can we need to extract public keys(for example if account is multisig, we need to notify all members). For that extractor locks(reader) cache two.
The problem can be, if between first and second locks, we started committing blocks. It means what writer can end his process, because he is waiting for end of first lock. But first lock can't finish, because he is waiting for end of second lock. But second lock waits for writer.

To avoid that, let's validate all transactions first, unlock cache, and after that notify about failures.